# Documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Supchat
  description: une API de chat pour une entreprise
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Serveur local Symfony
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Channels:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 3
          maxLength: 255
        workspace:
          $ref: '#/components/schemas/Workspaces'
        status:
          type: boolean
          description: true = public, false = private
        minRole:
          type: integer
          default: 1
          description: 1 = membre, 2 = modérateur, 3 = admin
      required: [name, workspace, status]
...
    Users:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          minLength: 3
          maxLength: 255
        lastName:
          type: string
          minLength: 3
          maxLength: 255
        userName:
          type: string
          minLength: 3
          maxLength: 255
        emailAddress:
          type: string
          format: email
          minLength: 3
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 255
          nullable: true
        role:
          type: string
          enum: ["ROLE_USER", "ROLE_ADMIN"]
        theme:
          type: boolean
        status:
          type: string
          minLength: 3
          maxLength: 255
      required: [firstName, lastName, userName, emailAddress, role, theme, status]
...
    Workspaces:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          minLength: 3
          maxLength: 255
        status:
          type: boolean
          description: true = public, false = private
        creator:
          $ref: '#/components/schemas/Users'
      required: [name, status, creator]
...
paths:
...
  /api/auth/register:
      post:
        tags: [Auth]
        summary: Inscription d’un nouvel utilisateur
        operationId: registerUser
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [firstName, lastName, userName, emailAddress, password]
                properties:
                  firstName:
                    type: string
                  lastName:
                    type: string
                  userName:
                    type: string
                  emailAddress:
                    type: string
                    format: email
                  password:
                    type: string
        responses:
          '201':
            description: Utilisateur inscrit avec succès
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                    userId:
                      type: integer
          '400':
            description: Données incomplètes
          '409':
            description: Email déjà utilisé
  /api/auth/login:
      post:
        tags: [Auth]
        summary: Connexion d’un utilisateur
        operationId: loginUser
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [emailAddress, password]
                properties:
                  emailAddress:
                    type: string
                    format: email
                  password:
                    type: string
        responses:
          '200':
            description: Connexion réussie
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    token:
                      type: string
                    user:
                      $ref: '#/components/schemas/Users'
          '400':
            description: Données incomplètes
          '401':
            description: Identifiants invalides
...
  /api/auth/refresh:
      get:
        tags: [Auth]
        summary: Rafraîchissement du token JWT
        operationId: refreshToken
        security:
          - BearerAuth: []
        responses:
          '200':
            description: Nouveau token généré
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    token:
                      type: string
          '401':
            description: Token invalide ou manquant
  /api/channels/{id}:
    get:
      tags: [Channels]
      summary: Récupère un channel
      operationId: getChannel
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails du channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channels'
        '403':
          description: Accès interdit
        '404':
          description: Channel introuvable

    put:
      tags: [Channels]
      summary: Met à jour un channel
      operationId: updateChannel
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: boolean
                min_role:
                  type: integer
      responses:
        '200':
          description: Canal mis à jour
        '400':
          description: Requête invalide

    delete:
      tags: [Channels]
      summary: Supprime un channel
      operationId: deleteChannel
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Canal supprimé
  /api/channels:
    post:
      tags: [Channels]
      summary: Crée un nouveau canal
      operationId: createChannel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, status]
              properties:
                workspace_id:
                  type: integer
                status:
                  type: boolean
                name:
                  type: string
                participants:
                  type: array
                  items:
                    type: integer
                min_role:
                  type: integer
      responses:
        '201':
          description: Canal créé
        '400':
          description: Données invalides ou incomplètes
        '403':
          description: Accès refusé
        '409':
          description: Canal déjà existant
  /api/channels/simple:
    post:
      tags: [Channels]
      summary: Crée une conversation privée simple
      operationId: createSimplePrivateChannel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [workspace_id, status, participants]
              properties:
                workspace_id:
                  type: integer
                status:
                  type: boolean
                participants:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Canal privé créé
        '400':
          description: Données invalides ou manquantes
        '409':
          description: Canal déjà existant
  /api/channels/{id}/privilege:
    post:
      tags: [Channels]
      summary: Récupère les privilèges d’un utilisateur dans un canal
      operationId: getChannelPrivilege
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required: [user_id]
              properties:
                user_id:
                  type: integer
      responses:
        '200':
          description: Privilèges renvoyés
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_admin:
                    type: boolean
                  can_moderate:
                    type: boolean
                  can_manage:
                    type: boolean
        '400':
          description: ID utilisateur manquant
        '404':
          description: Utilisateur introuvable
  /api/channels/by-name/{name}:
    get:
      tags: [Channels]
      summary: Récupère un canal par son nom
      operationId: getChannelByName
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Canal trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channels'
        '404':
          description: Canal non trouvé
  /api/workspaces/{id}/channels:
    get:
      tags: [Channels]
      summary: Liste les channels d’un workspace
      operationId: getChannelsByWorkspace
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channels'
...